fix test dir

+++ tests/shared.ts
@@ -3,8 +3,10 @@ import * as fs from "https://deno.land/std@0.205.0/fs/mod.ts";
 export const TEST_ENV_DIR = `.test-env`;
 
 export function init(): void {
-  fs.emptyDirSync(TEST_ENV_DIR);
-  Deno.chdir(TEST_ENV_DIR);
+  if (!Deno.cwd().endsWith(TEST_ENV_DIR)) {
+    fs.emptyDirSync(TEST_ENV_DIR);
+    Deno.chdir(TEST_ENV_DIR);
+  }
 }
 
 export function cleanup(rootDir: string): void {
